+++ src/containers/SoundsNavContainer.tsx
@@ -1,19 +1,8 @@
-import * as React from "react";
 import { connect } from "react-redux";
 import { getChosenTag, getTagsByOrder } from "../store/selectors";
 import { AppDispatch, chooseTag } from "../store/actions";
 import AppState from "../store/state";
 import SoundsNav, { Props } from "../components/SoundsNav";
-
-function SoundsNavContainer({
-  changeTag,
-  currentTag,
-  tags
-}: Props): React.ReactElement<Props> {
-  return (
-    <SoundsNav tags={tags} currentTag={currentTag} changeTag={changeTag} />
-  );
-}
 
 const mapStateToProps = (state: AppState) => ({
   currentTag: getChosenTag(state),
@@ -29,4 +18,4 @@ const mapDispatchToProps = (dispatch: AppDispatch) => ({
 export default connect<Props, {}, {}, AppState>(
   mapStateToProps,
   mapDispatchToProps
-)(SoundsNavContainer);
+)(SoundsNav);